[Feat] 화면 리렌더링 문제 해결 - 버튼만 렌더링
