+++ src/pages/teamain/StartButton.tsx
@@ -1,13 +1,46 @@
-import React from 'react';
+import { useEffect, useState } from 'react';
 import styles from '@/css/teamain/TeamMain.module.scss';
 
 interface StartButtonProps {
-  isActive: boolean;
-  timeLeft: number;
+  startTime: string;
   onClick: () => void;
 }
 
-function StartButton({ isActive, timeLeft, onClick }: StartButtonProps) {
+export default function StartButton({ startTime, onClick }: StartButtonProps) {
+  const [timeLeft, setTimeLeft] = useState(0);
+  const [isActive, setIsActive] = useState(false);
+
+  useEffect(() => {
+    if (!startTime) return;
+
+    const [hour, minute] = startTime.split(':').map(Number);
+    const today = new Date();
+    const startTimestamp = new Date(
+      today.getFullYear(),
+      today.getMonth(),
+      today.getDate(),
+      hour,
+      minute,
+      0
+    ).getTime();
+
+    const updateTimer = () => {
+      const now = Date.now();
+      const diff = Math.floor((startTimestamp - now) / 1000);
+      setTimeLeft(diff);
+
+      if (diff <= 1800 && diff > 0) setIsActive(true);
+      else if (diff <= 0) {
+        setIsActive(false);
+      }
+    };
+
+    updateTimer(); // 초기 실행
+    const timerId = setInterval(updateTimer, 1000);
+
+    return () => clearInterval(timerId);
+  }, [startTime]);
+
   const formatTime = (seconds: number) => {
     const min = Math.floor(seconds / 60).toString().padStart(2, '0');
     const sec = (seconds % 60).toString().padStart(2, '0');
@@ -26,5 +59,3 @@ function StartButton({ isActive, timeLeft, onClick }: StartButtonProps) {
     </button>
   );
 }
-
-export default React.memo(StartButton);
